+++ NEWS.md
@@ -1,5 +1,6 @@
 # compareGroups 4.9.1
 - bug fixed in cGroupsWUI()
+- bug fixed in compareGroups.fit when some variables cannot be described
 
 # compareGroups 4.9.0
 - new argument `nmax.method` to compute the N shown in the header. By default N is computed as the number of rows with at least one valid value in the described variables, but also N can be computed as the total number of rows in the data set or in each group regardless the available data in the described variables.